Change 'options' argument to 'params'
Previously the generator function took a final argument called
'options', which was the command lin eoptions dict from optparse. With
this change, that argument becomes 'params', which is a dict with
these keys:

options: the old options structure
build_files: the list of build files passed on the command line
generator_flags: a list of strings. These strings are taken from the
  --generator-flags command line argument. For example, if the command
  line includes "--generator-flags=opt,foo,bar" then this list will be
  ['opt', 'foo', 'bar']



git-svn-id: https://gyp.googlecode.com/svn/trunk@372 78cadc50-ecff-11dd-a971-7dbc132099af