Flesh out loadable_module support:
In parsing the GYP files, loadable_module counts as something that can't be
linked against and is a final linkable type (a deliverable).

Add a mapping for the msvc generator (matches shared_libary).

Add mappings in the xcode generator to handle both bundled and unbundled
loadable_modules.

Review URL: http://codereview.chromium.org/42464

git-svn-id: https://gyp.googlecode.com/svn/trunk@386 78cadc50-ecff-11dd-a971-7dbc132099af