make: test for ar's "T" flag at gyp time
1) Add a "system_test" module to gyp for running tests against
   the current system.
2) Use this module to test for an optional build flag.

I intend to add more tests in the future, so this sketches out
the way I intend for it to work.

The test (which compiles and links a test program) is fast
('python system_test.py' takes 135ms locally) so it doesn't
slow down gyp in any significant way.

Review URL: http://codereview.chromium.org/6483003

git-svn-id: https://gyp.googlecode.com/svn/trunk@884 78cadc50-ecff-11dd-a971-7dbc132099af